Added colour picking
Basically will render scene in colour codes - use glGetPixels to grab the colour ID on a mouse click to see if something's being picked. Then the buffer clears and the scene renders as normal - buffer's don't swap, so none of the colour codes are shown. Buggy as fuck though -- need to figure out what's up.